Update ui package
ChashUi converted to abstract class to prep for gui support
Extracted ChashUi logic into ChashConsole that extends ChashUi
